updated to v1.5.5f

+++ zc_install/includes/classes/class.zcDatabaseInstaller.php
@@ -2,9 +2,9 @@
 /**
  * file contains zcDatabaseInstaller Class
  * @package Installer
- * @copyright Copyright 2003-2015 Zen Cart Development Team
+ * @copyright Copyright 2003-2016 Zen Cart Development Team
  * @license http://www.zen-cart.com/license/2_0.txt GNU Public License V2.0
- * @version $Id: New in v1.6.0
+ * @version $Id: Author: DrByte  Thu Mar 21 16:18:21 2016 -0500 New in v1.5.5 $
  *
  */
 /**
@@ -32,17 +32,17 @@ class zcDatabaseInstaller
     $this->dbPassword = $options['db_password'];
     $this->dbName = $options['db_name'];
     $this->dbPrefix = $options['db_prefix'];
-    $this->dbCharset = $options['db_charset'];
+    $this->dbCharset = trim($options['db_charset']) == '' ? 'utf8' : $options['db_charset'];
     $this->dbType = in_array($options['db_type'], $dbtypes) ? $options['db_type'] : 'mysql';
     $this->dieOnErrors = isset($options['dieOnErrors']) ? (bool)$options['dieOnErrors'] : FALSE;
     $this->errors = array();
     $this->basicParseStrings = array(
-    'DROP TABLE IF EXISTS ',
+    'DROP TABLE ',
     'CREATE TABLE ',
     'REPLACE INTO ',
     'INSERT INTO ',
     'INSERT IGNORE INTO ',
-    'ALTER IGNORE TABLE ',
+//    'ALTER IGNORE TABLE ',
     'ALTER TABLE ',
     'TRUNCATE TABLE ',
     'RENAME TABLE ',
@@ -103,6 +103,7 @@ class zcDatabaseInstaller
       if ( substr($this->line,-1) ==  ';')
       {
         if (substr($this->newLine,-1)==' ') $this->newLine = substr($this->newLine,0,(strlen($this->newLine)-1));
+        if (substr($this->newLine,-1)==')') $this->newLine = substr($this->newLine,0,(strlen($this->newLine)-1)) . ' )';
         $this->keepTogetherCount++;
         if ($this->keepTogetherCount == $this->keepTogetherLines)
         {
@@ -157,15 +158,29 @@ class zcDatabaseInstaller
 //    $this->writeUpgradeExceptions($this->line, '', $this->sqlFile);
 //    logDetails($sql, $this->sqlFile);
     $result = $this->db->execute($sql);
-    if (!$result)
+    if (!$result || $result->link->errno != 0)
     {
-      //echo $this->db->errorText;
-      $this->writeUpgradeExceptions($this->line, $this->db->error_text, $this->sqlFile);
+      $this->writeUpgradeExceptions($this->line, $this->db->error_number . ': ' . $this->db->error_text);
+      error_log("MySQL error " . $this->db->error_number . " encountered during zc_install:\n" . $this->db->error_text . "\n" . $this->line . "\n---------------\n\n");
     }
   }
-  public function parserDropTableIfExists ()
+  public function parserDropTable()
   {
-    $this->line = 'DROP TABLE IF EXISTS ' . $this->dbPrefix . substr($this->line, 21);
+    $table = (strtoupper($this->lineSplit[2].' '.$this->lineSplit[3]) == 'IF EXISTS') ? $this->lineSplit[4] : $this->lineSplit[2];
+
+    if (!$this->tableExists($table) && (strtoupper($this->lineSplit[2].' '.$this->lineSplit[3]) != 'IF EXISTS'))
+    {
+      $result = sprintf(REASON_TABLE_NOT_FOUND, $table).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else {
+      if (strtoupper($this->lineSplit[2].' '.$this->lineSplit[3]) != 'IF EXISTS')
+      {
+        $this->line = 'DROP TABLE ' . $this->dbPrefix . substr($this->line, 11);
+      } else {
+        $this->line = 'DROP TABLE IF EXISTS ' . $this->dbPrefix . substr($this->line, 21);
+      }
+    }
   }
   public function parserCreateTable()
   {
@@ -196,6 +211,54 @@ class zcDatabaseInstaller
     } else
     {
       $this->line = 'INSERT INTO ' . $this->dbPrefix . substr($this->line, 12);
+    }
+  }
+  public function parserInsertIgnoreInto()
+  {
+    if (!$this->tableExists($this->lineSplit[3]))
+    {
+      if (!isset($result)) $result = sprintf(REASON_TABLE_NOT_FOUND, $this->lineSplit[3]).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else
+    {
+      $this->line = 'INSERT IGNORE INTO ' . $this->dbPrefix . substr($this->line, 19);
+    }
+  }
+  public function parserTruncateTable()
+  {
+    if (!$this->tableExists($this->lineSplit[2]))
+    {
+      if (!isset($result)) $result = sprintf(REASON_TABLE_NOT_FOUND, $this->lineSplit[2]).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else
+    {
+      $this->line = 'TRUNCATE TABLE ' . $this->dbPrefix . substr($this->line, 15);
+    }
+  }
+  public function parserFrom()
+  {
+    if (!$this->tableExists($this->lineSplit[1]))
+    {
+      if (!isset($result)) $result = sprintf(REASON_TABLE_NOT_FOUND, $this->lineSplit[1]).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else
+    {
+      $this->line = 'FROM ' . $this->dbPrefix . substr($this->line, 5);
+    }
+  }
+  public function parserDeleteFrom()
+  {
+    if (!$this->tableExists($this->lineSplit[2]))
+    {
+      if (!isset($result)) $result = sprintf(REASON_TABLE_NOT_FOUND, $this->lineSplit[2]).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else
+    {
+      $this->line = 'DELETE FROM ' . $this->dbPrefix . substr($this->line, 12);
     }
   }
   public function parserReplaceInto()
@@ -280,12 +343,30 @@ class zcDatabaseInstaller
       }
     }
   }
-  public function writeUpgradeExceptions($line, $message, $sqlFile)
+  public function parserRenameTable()
+  {
+    if (!$this->tableExists($this->lineSplit[2])) 
+    {
+      if (!isset($result)) $result = sprintf(REASON_TABLE_NOT_FOUND, $table).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else {
+      if ($this->tableExists($this->lineSplit[4]))
+      {
+        if (!isset($result)) $result = sprintf(REASON_TABLE_ALREADY_EXISTS, $table);
+        $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+        $this->ignoreLine = true;
+      } else {
+        $this->line = 'RENAME TABLE ' . $this->dbPrefix . $this->lineSplit[2] . ' TO ' . $this->dbPrefix . substr($this->line, (13 + strlen($this->lineSplit[2]) + 4));
+      }
+    }
+  }
+  public function writeUpgradeExceptions($line, $message, $sqlFile = '')
   {
     logDetails($line . '  ' . $message . '  ' . $sqlFile, 'upgradeException');
     $this->upgradeExceptions[] = $message;
     $this->createExceptionsTable();
-    $sql="INSERT INTO " . $this->dbPrefix . TABLE_UPGRADE_EXCEPTIONS . " VALUES (0,:file:, :reason:, now(), :line:)";
+    $sql="INSERT INTO " . $this->dbPrefix . TABLE_UPGRADE_EXCEPTIONS . " (sql_file, reason, errordate, sqlstatement) VALUES (:file:, :reason:, now(), :line:)";
     $sql = $this->db->bindVars($sql, ':file:', $sqlFile, 'string');
     $sql = $this->db->bindVars($sql, ':reason:', $message, 'string');
     $sql = $this->db->bindVars($sql, ':line:', $line, 'string');
@@ -361,6 +442,7 @@ class zcDatabaseInstaller
   }
   public function tableColumnExists($table, $column)
   {
+    if (!defined('DB_DATABASE')) define('DB_DATABASE', $this->dbName);
     $check = $this->db->Execute(
       'SHOW COLUMNS FROM `' . DB_DATABASE . '`.`' . $this->db->prepare_input($table) . '` ' .
       'WHERE `Field` = \'' . $this->db->prepare_input($column) . '\''
@@ -369,6 +451,7 @@ class zcDatabaseInstaller
   }
   public function tableIndexExists($table, $index)
   {
+    if (!defined('DB_DATABASE')) define('DB_DATABASE', $this->dbName);
     $check = $this->db->Execute(
       'SHOW INDEX FROM `' . DB_DATABASE . '`.`' . $this->db->prepare_input($table) . '` ' .
       'WHERE `Key_name` = \'' . $this->db->prepare_input($index) . '\''
